Network_C_server/Code/loadtests/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
gevent==1.3.7
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
locustio==0.9.0
MarkupSafe==1.1.0
msgpack==0.6.0
pyzmq==17.1.2
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
